# Conflicts:
#	src/controllers/AssetsController.php

+++ src/elements/actions/Delete.php
@@ -174,13 +174,17 @@ JS, [static::class]);
         $user = Craft::$app->getUser()->getIdentity();
 
         foreach ($query->all() as $element) {
-            if (!$element->canDelete($user)) {
+            if (!$elementsService->canView($element, $user) || !$elementsService->canDelete($element, $user)) {
                 continue;
             }
             if (!isset($deletedElementIds[$element->id])) {
                 if ($withDescendants) {
                     foreach ($element->getDescendants()->all() as $descendant) {
-                        if (!isset($deletedElementIds[$descendant->id]) && $descendant->canDelete($user)) {
+                        if (
+                            !isset($deletedElementIds[$descendant->id]) &&
+                            $elementsService->canView($descendant, $user) &&
+                            $elementsService->canDelete($descendant, $user)
+                        ) {
                             $elementsService->deleteElement($descendant);
                             $deletedElementIds[$descendant->id] = true;
                         }
